routing: solve interfaces not being casted into structures correctly
I honestly don't get how this could be working on the same Go version I've been testing on, and then all of a sudden, out of the blue, it just stops working. It makes no sense at all. Casting a map[string]interface{} type annotated with JSON hints should work automatically and just fine right? I'm not crazy, this did work, didn't it? Whatever the case now I get to enjoy rewriting a bunch of sections that used to work, and solve problems that didn't exist until Go just invented them while I was sitting doing my work.

Before, authenticated API requests would only 0ms on my local machine, which is very fast to say the least. Now with this new solution, a 16ms performance penalty is incured, and I had to introduce yet another dependency that has to be compiled, resulting in a larger binary size. That may not seem like a lot, but in a production environment where latency over the network, memory pressure, other processes taking CPU time, and database queries, this number could easily be inflated, and now I've inflated it even more. One day in the future I'll be able to get to the bottom of this issue, maybe Google will have even solved this problem. Maybe I'm just too angry, but this issue did take up around 8 hours of my time, just for me to have to default on a inferior solution simply because "magic", fun.